Started on packet analysis mode, player-gc can be disabled, more config
Packet analysis will aid in understanding the mc protocol (not finished, 
doesn't work as of yet but probably will later).

Got rid of the "safe" on the end of functions it was temporary for 
testing mutexes and preventing data races (still need to think about 
semaphores).

Also realised that NOT-ing the negative numbers may be more efficient 
than multiplying them.

Started working on packets been able to created and used later (futures)